pyproject.toml
[project]
name = "proffespor-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.13",
    "django-jazzmin>=3.0.4",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
]
